Torn signature v0.2
Added caching of images.  thiis way each image created is cached for 10 seconds, thus reducing requests, on load. (now with RW posts on forums, it occured that due to multiple replies, the request to the API timed out. This solves the issue with the drawback that the image might be as much as 9 seconds old.